Codegen for DoubleArrays.

+++ src/main/java/edu/jhu/prim/arrays/DoubleArrays.java
@@ -435,7 +435,7 @@ public class DoubleArrays {
     /** Checks in O(n) by linear search if the array contains the value. */
     public static boolean contains(double[][] array, double value, double delta) {
         for (int i=0; i<array.length; i++) {
-            if (contains(array[i], value, 1e-13)) {
+            if (contains(array[i], value, (double) 1e-13)) {
                 return true;
             }
         }
